low-level/requirements.txt
jsonschema==3.1.1
pika==1.1.0
pyserial==3.4
python-daemon==2.2.4
requests==2.22.0
zope.component==4.6
zope.event==4.4
zope.interface==4.7.1
pyinotify==0.9.4
